♻️ refactor: colocate provider model id helpers
