+++ packages/model-runtime/src/providers/anthropic/claudeThinkingHistory.ts
@@ -1,5 +1,5 @@
 import type { ChatStreamPayload, OpenAIChatMessage, UserMessageContentPart } from '../../types';
-import { isThinkingWithToolClaudeModel } from '../../utils/claudeModelId';
+import { isThinkingWithToolClaudeModel } from './claudeModelId';
 
 interface ClaudeThinkingHistoryMessageMetadata {
   model?: null | string;
